Move Instance isBelongsToSSA method to PrismLogic
Adam explains it better than me:
"I try very hard to avoid mixing logic with data structure definitions
with my own code. The only time I allow myself exceptions is when there
is a property that is inherently a part of the object itself, which is
to say it naturally arises as a part of the instance itself. It follows
that this is only possible if it does't take any parameters.
One rule of thumb is that you shouldn't need any tests for the models
package, which makes sense if it's just modelling data."
